added authentication to routes

+++ app/Http/Controllers/WorkoutController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 use App\Workout;
-
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class WorkoutController extends Controller
@@ -26,7 +26,12 @@ class WorkoutController extends Controller
      */
     public function create()
     {
-        return view('workouts.create');
+        if (Auth::check()) {
+            return view('workouts.create');
+        } else {
+            return redirect('/login');
+        }
+        
     }
 
     /**
@@ -37,18 +42,22 @@ class WorkoutController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'reps'=>'required',
-            'exercise'=>'required'
-        ]);
+        if(Auth::check()){
+            $request->validate([
+                'reps'=>'required',
+                'exercise'=>'required'
+            ]);
 
-        $workout = new Workout([
-            'reps' => $request->get('reps'),
-            'exercise'=> $request->get('exercise')
-        ]);
+            $workout = new Workout([
+                'reps' => $request->get('reps'),
+                'exercise'=> $request->get('exercise')
+            ]);
 
-        $workout->save();
-        return redirect('/workouts')->with('success', 'Workout saved!');
+            $workout->save();
+            return redirect('/workouts')->with('success', 'Workout saved!');
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -70,8 +79,14 @@ class WorkoutController extends Controller
      */
     public function edit($id)
     {
-        $workout = Workout::find($id);
-        return view('workouts.edit', compact('workout'));
+        if(Auth::check()){
+
+            $workout = Workout::find($id);
+            return view('workouts.edit', compact('workout'));
+
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -83,18 +98,22 @@ class WorkoutController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'reps'=> 'required',
-            'exercise'=>'required'
-        ]);
+        if(Auth::check()){
+            $request->validate([
+                'reps'=> 'required',
+                'exercise'=>'required'
+            ]);
 
-        $workout = Workout::find($id);
-        $workout->reps = $request->get('reps');
-        $workout->exercise = $request->get('exercise');
+            $workout = Workout::find($id);
+            $workout->reps = $request->get('reps');
+            $workout->exercise = $request->get('exercise');
 
-        $workout->save();
+            $workout->save();
 
-        return redirect('/workouts')->with('success', 'Workout updated!');
+            return redirect('/workouts')->with('success', 'Workout updated!');
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -105,9 +124,14 @@ class WorkoutController extends Controller
      */
     public function destroy($id)
     {
-        $workout = Workout::find($id);
-        $workout->delete();
-
-        return redirect('/workouts')->with('success', 'Workout deleted!');
+        if(Auth::check()){
+            $workout = Workout::find($id);
+            $workout->delete();
+    
+            return redirect('/workouts')->with('success', 'Workout deleted!');
+        } else {
+            
+            return redirect('/login');
+        }
     }
 }
